Moved Hassium.Parser (the class) to Hassium.Parser (the namespace) so now it's Hassium.Parser.Parser (it was causing build problems)
