+++ src/Hassium/Hassium.csproj
@@ -42,6 +42,7 @@
   <ItemGroup>
     <Compile Include="MyClass.cs" />
     <Compile Include="Parser\Ast\ArrayGetNode.cs" />
+    <Compile Include="Parser\Ast\ArrayInitializerNode.cs" />
     <Compile Include="Parser\Ast\ForEachNode.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Lexer\IFunction.cs" />
@@ -87,7 +88,5 @@
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup />
-  <ItemGroup>
-    <Folder Include="Semantics\" />
-  </ItemGroup>
+  <ItemGroup />
 </Project>
