+++ src/Hassium/Parser/Ast/ArrayInitializerNode.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Hassium.Parser.Ast
+{
+    class ArrayInitializerNode
+    {
+    }
+}
