+++ src/Hassium/Parser/Ast/ForNode.cs
@@ -44,7 +44,7 @@ namespace Hassium
             this.Children.Add(body);
         }
 
-        public static AstNode Parse(Parser parser)
+        public static AstNode Parse(Parser.Parser parser)
         {
             parser.ExpectToken(TokenType.Identifier, "for");
             parser.ExpectToken(TokenType.Parentheses, "(");
